+++ Metrics/Graphite/GraphiteExtensions.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Configuration;
 
-using Metrics.Logging;
 using Metrics.Reporters;
 
 namespace Metrics.Graphite
@@ -47,33 +45,5 @@ namespace Metrics.Graphite
         {
             return reports.WithReport(new GraphiteReport(graphiteLink), interval);
         }
-
-        public static MetricsReports WithGraphiteFromConfig(this MetricsReports reports)
-        {
-            var graphiteMetricsUri = ConfigurationManager.AppSettings["Metrics.Graphite.Uri"];
-            var graphiteMetricsInterval = ConfigurationManager.AppSettings["Metrics.Graphite.Interval.Seconds"];
-
-            if (!string.IsNullOrEmpty(graphiteMetricsUri) && !string.IsNullOrEmpty(graphiteMetricsInterval))
-            {
-                Uri uri;
-                if (!Uri.TryCreate(graphiteMetricsUri, UriKind.Absolute, out uri))
-                {
-                    throw new Exception("Invalid Metrics Configuration: Metrics.Graphite.Uri must be a valid absolute URI");
-                }
-                int seconds;
-                if (!int.TryParse(graphiteMetricsInterval, out seconds) || seconds <= 0)
-                {
-                    throw new Exception("Invalid Metrics Configuration: Metrics.Graphite.Interval.Seconds Seconds must be an integer > 0");
-                }
-
-                log.Debug(() => $"Metrics: Sending Graphite reports to {uri} every {seconds} seconds.");
-                return reports.WithGraphite(uri, TimeSpan.FromSeconds(seconds));
-            }
-
-            log.Info(() => "Graphite configuration missing from config file. Graphite reporting will be disabled.");
-            return reports;
-        }
-
-        private static readonly ILog log = LogProvider.GetCurrentClassLogger();
     }
 }